refactor toast notifications to remove theme dependency and streamline usage
