+++ src/components/layout/Root.jsx
@@ -3,8 +3,12 @@ import Box from "@mui/material/Box";
 import Header from "./Header.jsx";
 import {Container} from "@mui/material";
 import Footer from "./Footer.jsx";
+import {ToastContainer} from 'react-toastify';
+import useTheme from "../../hooks/useTheme.js";
+
 
 export default function Root() {
+  const currentTheme = useTheme();
   return (
     <Container sx={{minHeight: '100vh', display: 'flex', flexDirection: 'column'}}>
       <Header/>
@@ -17,6 +21,11 @@ export default function Root() {
       >
         <Outlet/>
       </Box>
+      <ToastContainer
+        position="bottom-right"
+        role="alert"
+        aria-live="polite"
+        theme={currentTheme === 'dark' ? 'dark' : 'light'}/>
       <Footer/>
     </Container>
   );
